refactor: Token helper functions extracted to different directory and combined both token endpoints files
